feat(router): reimplement core routing functionality with context and guard extensions
